fix: move `deriveKeyBitsFromPassword` to `jwk` utils
